backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
deprecation==2.1.0
dill==0.4.0
ecdsa==0.19.1
fastapi==0.122.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
joblib==1.5.2
multidict==6.7.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
postgrest==2.24.0
prettytable==3.17.0
propcache==0.4.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
realtime==2.24.0
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
skops==0.13.0
sniffio==1.3.1
starlette==0.50.0
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
